bGems, Index, Main, Side, and Shrine Quests, Towers
Added fabrics table to database for Paraglider Fabrics, and display the fabrics collected through chests, quests, or pony points in a table on the homepage.
Added Bubbul Gems page - clone of Pony Points page, listing Koltins shop and items to buy with gems.
Shrines and Shrine Quests updated - the name of the quest or the shrine is now a link to take the user to the corresponding entry on the opposite page.
Main and Side Quests updated, working on functionalty to show sub quests. Works for Side quests like "Photographing a Chuchu" where the sub quest is in the subquests table, but issues listing subquests like "A Picture For..." where the subquest is another quest in the quests table.