Armors Mobile and Javascript
changed Armors page technicalities - due to an issue showing the tables properly, instead of tables it is now using a plethora of divs with id's.
updated Armor toggle functionality to properly show upgrade divs based on armor collected and number of fairies. Need to add specification for armor sets that do not upgrade. Also need to add single armor pieces.